requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
